Create Study MBBS from Best Medical College in Russia 2021-22  Twinkle InstituteAB
Before getting to know about doing MBBS in Russia let us know something about the most preferred destination Russia and why you need to study MBBS in Russia Along with a lot of advantages and disadvantages of pursuing MBBS in Russia.

Medical education from Top Medical College in Russia is a golden opportunity for Indian medical aspirants. According to WHO, Russian medical universities in top-ranking take approximately 30 positions in the list. Medical universities in Russia have significantly lower fees as compared to other countries. There are more than 50 Best Universities in Russia for MBBS, which is enlisted in MCI. Everyone here in India is requesting the Top Universities in Russia for MBBS. MBBS in Russia You can find out the many low-cost fee medical universities.

If you have scored low in your NEET and you did not get a medical seat in India, then you have plenty of opportunities to get medical education from the best university of medical from abroad, you can fulfil your dreams of becoming a doctor. We provide the best educational counselling and guidelines for a different country like MBBS In Russia, China, Ukraine, Kazakhstan, Kyrgyzstan, More. We have been providing educational counselling for the past few years, and we have sent more than 100 Indian medical students abroad.

Russia is home to some of the oldest and most respected medical universities in the world, Russia has been attracting students from different countries for the past three decades. MBBS is one of the most reputed courses for which young people flock from every corner of the world, every year. The education department of Russia is generally high advanced and it lies among the most literate countries of the world with a literacy rate of 99%. The significance is that the government lays on science and technology has also led to the growth and maintenance of high technology medical universities in Russia. The signing of the Bologna agreement made things even better for all the international student visitors or applicants to Russia since the standard and kind of medical study provided in the Russia country began to be at standard with that in the rest of whole Europe. MBBS degrees awarded by medical universities in Russia were also to be accepted everywhere else in Europe and also in other countries of the world. The main boon of studying MBBS in Russia is the low-cost accommodation, recognition of degrees by WHO, MCI and other international medical bodies and high quality of medical education. There is several doctors, medical staffs and candidates in Russia. If we put an estimation that would be around ninety lakhs. Indian medical candidates who complete their MBBS from Russia can also pursue their degree of masters and further research studies here because the scope in the field of medicine is quite good in this country. Students can also choose to return to their native countries and can practice medicines as professionals. To practice in India students must clear the MCI screening test (FMGE) to practice doctorate.

The duration of the whole MBBS course in Russia is five years and eight months which includes the year of the preparatory course of medicines. MBBS in Russia introduces students to the Russian language and culture through reading and writing and communicating with them which thus aims to help them in communication during their stay and the medicinal practice in hospitals in that country. It is also provided to lessen the gaps that might have occurred in a student’s school primary and secondary education, especially in their science knowledge. This also creates a connection between the science studied at the school level in their home country and the kind they are about to begin studying in Russia, you will clearly see the difference.

Russian medical universities do not have any entrance examination or requirements for admission in MBBS in Russia. There is no requirement of paying huge donations during admission. Their MBBS course syllabus and standards are all superintended by the Ministry of Education and Science. Students are provided with both the option of studying MBBS in English and Russian language. The average yearly university fee of Russian medical universities is around 6000 USD which is much less as compared to any other medical universities in a European country. The whole cost of living and accommodation can be kept between 100 USD to 150 USD and it can be above according to the student’s expense. Most medical colleges provide hostel facilities to the students and rented apartments are also located nearby the universities.

Reasons why you should pursue your MBBS from Russia:-

Affordable fees to the best MBBS course structure.
Russia being the best in academics along with sports and other co-curricular activities.
It is mandatory to be 17 years Age not less than that to be eligible for applying.
You just need to score 50% in your 12th standard and some cases but for some medical universities in Russia, you may be required to have at least 60% in 12th boards.
There is a requirement to be qualified in NEET Exam to be eligible to practice medicine in India which is a great benefit and hassle-free.
Scholarship will be provided to international medical students who are eligible and score well.
All medical universities in Russia are Government organizations and degrees are recognized globally and accredited by the WHO & Medical Council of India, which makes students eligible to pursue an education in any country