Create Why you should pursue MBBS from Russia 2021-22? Twinkle InstituteAB
Russia is a country situated in Eastern Europe. It became a popular country for higher medical education and one of the popular destinations for many MBBS candidates who want to study MBBS in Russia Medical College.

MBBS in Russia is now recognized worldwide. There are many top government medical universities in Russia that are approved by the Medical Council of India (MCI). Medical universities in Russia offer the 6-year MBBS at a low and affordable fee cost. They offer high-quality medical education to both local and international medical students in the English language. Nowadays Medicine in Russia is popular among the Indian medical aspirants as there are many top medical universities that offer MBBS in Europe and are actually listed in the world’s famous medical bodies like WHO and MCI. This, in turn, ensures the students practice medicine anywhere after completing their MBBS degree in Russia. The program of Study in Russia for MBBS can be completed in five years and eight months. In these almost 6 years, candidates are involved in many co-curricular activities apart from their academics and theoretical/classroom sessions. To make learning more interesting and interactive, these best medical universities in Russia enable the students to be involved in lab research fields and other live projects so that students can gain more practical knowledge. Students are also made to attend clinical training with the assistance of medical professionals to improve their practical knowledge and skills. In parallel to the academic activities, these top Medical Universities in Russia arrange extracurricular activities for the students to get them involved in sports activities and different cultural fests. Every medical university has its own admission procedure. To get admission into these top medical universities in Russia, medical candidates must appear and clear NEET entrance examinations to take place in India. NEET is a mandatory examination for all Indian citizens and overseas citizens of India who want to study MBBS in India as well as abroad.

Once you have appeared for the NEET, then the score is valid for up to three years of studying MBBS from abroad.

After qualifying NEET, medical aspirants can get direct admission on the basis of the score for MBBS in Russia for Indians. Higher Education in the field of medicine is regulated by the Ministry of Education and Science of Russia. MBBS in Russia includes both academic as well as professional qualifications. Therefore, the Medical Universities in Russia are also recognized by the health Ministry of Education and other medical organisations of the world. Along with that, the top medical universities in Russia are approved by the both Medical Council of India (MCI) and the Medical Council of Canada (MCC). They are also verified by the Educational Commission for Foreign Medical Graduates (ECFMG). These medical universities are being registered with the Foundation for Advancement of International Medical Education and Research (FAIMER) and are listed with the World Directory of Medical Schools (WDOMS). After graduating from MBBS college in Russia, medical students can practice medicine in India and other countries worldwide after clearing the Medical Licensing Examination or MCI screening test.

Students after completing their MBBS course can also practice medicine in any country of the world subject to clearing the country’s particular Medical Licensing Examination.  All medical universities in Russia provide the best medical education to all its local and international students. Every medical university is fully equipped with advanced technology to provide the advanced medical program to the students studying there, pursuing MBBS abroad at a low cost teach the students to manage all their expense requirements within their budget and it is not costly living up there. Those students studying MBBS in Russia get plenty of opportunities to learn so many new things during their whole academic years. The top 50 medical universities are the cheapest medical universities in Russia that are offering MBBS programs at an effective cost. Out of these, Top 30 medical colleges in Russia are approved by the Medical Council of India (MCI). These universities are fully provided with modern technologies to infuse modern learning to the students studying MBBS in Russia.

There are many people who think that the MBBS course costs the same in India and out of India. So many have the idea that the entire expenses of studying MBBS abroad will be the same as spending in India for this course. However, people are unaware of so many things about MBBS abroad that we will be introducing it to you. MBBS in Russia costs way cheaper and affordable as compared to India.

On top of that, the subjects that are being taught in Russian medical universities are the same as the subjects taught in India. Many of the medical universities in Russia offering MBBS to Indian students as well as to other foreign medical aspirants are English Medium medical universities. Russia medical study is popularly known for its worldwide acceptance, affordable tuition fees, low cost of living, accommodation etc. Thus, many international students move to Russia for higher medicine study. There are so many reasons classified above that why a medical aspirant should opt for MBBS in Russia.