Create Get your Medical Degree from Russian Medical Universities 2021-22 Twinkle InstituteAB
MBBS in Russia has always been the most preferred choice for Indian medical students in India. There are more than 10,000 Indian based medical aspirants that are currently pursuing MBBS in Russia in about 50+ MCI recognized Top ten medical university In Russia all over Russia. Every Russia Medical College is being listed and recognized in WHO (World Health Organization) list and also recognized by the Medical Council of India (MCI).

Between the year 2012 to 2017, more than 11000 students have come back to India after clearing FMGE, which clears states that more than 2,000 Indian medical students(on average) start their Medical Practices after completing MBBS from Russia.

MBBS in Russia being one of the most affordable destinations available for studying MBBS, thousands of Indian students make it a choice every year. MBBS Course in Russia is considered one of the most advanced amongst today’s peers and uses the latest and high-quality teaching methods and Techniques in the Best MBBS college in Russia. Also, considering the healthy relations between Indian and Russia, the Russian government has always made Indian students feel welcomed in their country. The natives of Russia are generally very cooperative and helpful towards Indians studying there.

MBBS in Russia focuses on wholesome and all-around development of the student’s knowledge in medicine as well as on the extensive development of the student in their chosen field of specialization. The Study in Russia for MBBS is amongst the best education in the world, with the student’s performance and progress measured against several parameters. The student is assured to become a successful doctor after returning to India. But it is compulsory provided the students should work hard and dedicate to give their 100% in studying. MBBS in Russia is already being pursued by thousands of Indian medical students who are satisfied with the high quality of education together with the low cost of tuition fee and affordable expense of living in Russia.

Although most of the MBBS courses are taught in the Russian Language still there are few medical universities in Russia which offer complete MBBS course in both English and Russian language which is an added advantage for the International students.

Russia has the most top universities, the level of these higher graduates universities in terms of percentage of the population in the world, so this is the reason why the literacy rate is up to 99% over there. The entrance to subsidized higher education is highly competitive over there.

As a result of the great significance of science and technology, Russian medical education, mathematical, scientific, and aerospace research is generally of high quality with modern methods of teaching.

Medical college in Russia are extremely recognized and hence the student can work anywhere in the world after completing their MBBS course in Russia.

The list of medical universities in Russia mentioned on our site are already recognized by the Medical Council of India where students can pursue and make their dreams of studying MBBS in Russia and the degree is valid all over the worldwide and students will be prepared for MCI screening test conducted in India every year for every MBBS graduate.

The idea of living in Russia seems expensive but actually, it is one of the best experiences of someone’s life. Being in such a beautiful and peaceful place is every student’s dream. Russia is not only the country of white people but the country of every individual, it is also the country of the good-hearted and educated people. As being the largest country in the world, there are many many things to explore for students in this country.

The population of

Russia, one of the biggest country in the world which has more than 4 crore and 50 lakhs population. Moscow is the capital of this beautiful and incredible country that have so many medical universities. The biggest city of Russia is none other than its capital that is Moscow. It is a very amazing fact to know that there are a total of nine time zones in the country Russia. It means that if there is a day at one corner of Russia then there will be a night at the opposite corner of Russia.

The whole campus of Russian medical universities is very well maintained and equipped with modern types of equipment and laboratory machinery. However, MBBS in Russia is very beneficial as most of the medical universities of Russia are recognized by WHO, UNESCO, MCI, and other world medical organisations.

As mentioned above, studying MBBS in Russia for Indian students is quite cheap as compared to any other country. It is not only the fees but in daily expenses, students can save much more as compared to other countries. It also varies and depends on the student’s lifestyle.